requirements.txt
openpyxl
requests
urllib3
python-dotenv
PyYAML
ansible